Move Inheritance module and classes to spec
