+++ spec/support/fixtures.rb
@@ -1850,73 +1850,6 @@ module Flash
   end
 end
 
-module Inheritance
-  class Action < Hanami::Action
-    before :log_base_action
-
-    private
-
-    def log_base_action(*, res)
-      res[:base_action] = true
-    end
-  end
-
-  class AuthenticatedAction < Action
-    before :authenticate!
-
-    private
-
-    def authenticate!(*, res)
-      res[:authenticated] = true
-    end
-  end
-
-  module Actions
-    module Books
-      class RestfulAction < AuthenticatedAction
-        before :find_book
-        after :render
-
-        private
-
-        def find_book(req, res)
-          res[:book] = "book #{req.params[:id]}"
-        end
-
-        def render(*, res)
-          res.body = res.exposures.keys
-        end
-      end
-
-      class Show < RestfulAction
-        def handle(*, res)
-          res[:found] = true
-        end
-      end
-
-      class Destroy < Show
-        def handle(*, res)
-          super
-          res[:destroyed] = true
-        end
-      end
-    end
-  end
-
-  class Application
-    def initialize
-      @routes = Hanami::Router.new do
-        get "/books/:id", to: Inheritance::Controllers::Books::Show.new
-        delete "/books/:id", to: Inheritance::Controllers::Books::Destroy.new
-      end
-    end
-
-    def call(env)
-      @routes.call(env)
-    end
-  end
-end
-
 class ContractActionBase < Hanami::Action
   def handle(request, response)
     if request.params.valid?
